MB-14637 scanner: Use new byte streaming row results protocol
We used to have protobuf messages containing an array of rows
for streaming scan results. But, it adds overhead of building
a list of rows in heap and also protobuf encoding time. But,
we could maximize the performance by performing zero copy
writes in the index pipeline writer component.

Migrated protobuf based row streaming to raw bytes streaming.
The packet layout is as follows:

[protobuf_header len=total_bytes][sk_len][sk][pk_len][pk]....[sk_len][sk][pk_len][pk]

Change-Id: Ia9d71742cf4c264e69f13f5c215e62ffacf709ad