MB-100 indexer: Make mutation sync pool configurable
It is noticed that mutation sync pool causes major scalability
bottleneck for memdb. CPU utilization jumps from 25 cores to 35 cores
once sync pool is disabled. Also it is noticed that using sync pool results
in high memory usage for the process. Hence it is better to disable sync pool
by default and enable it for indexer with small number of CPUs using settings.

Change-Id: Ie7bd92bdf5ab096ba632448f05a333ba623539b7